requirements.txt
Django
gunicorn
whitenoise
django-rest-framework
requests
django-sslserver
